Init
Ran the following prompt on ChatGPT:

```
Act as a professional software engineer, that is aware of fast execution times, and most of all - readability.

In C#, create a static method that receives 3 parameters:

* "backgroundColor" (think what are the best practices to handle giving a color as a parameter)
* (optional) "text" (string) -> by default it is null.
* (optional) "edgeLength" (integer) -> by default it is set to 500 pixels.

The method will generate an new PNG / SVG, where:

* The "backgroundColor" is the background color of the picture.
* The "text" is optional, and should be the foreground text on-top of the background color, that is aligned to the center of the picture.
* The "edgeLength" being the edge length in pixels of the squared picture.

If that makes sense, you can see references online on companies that does this "default picture generation" method as like done in Atlassian Jira (exactly of how we want it) or GitHub / GitLab is some kind of similar.

In business perspective, the motive behind this method is that a new user could have a default picture generated for him even if he did not upload a picture already. So the background color could be picked by the user or randomly, and the text would be his name's first letters (for example: John Doe would be JD).

Finally, the method should be generating a PNG/SVG picture in the local filesystem as a file, and then output the path to the file as the output of the method.
```

The outputs of `DefaultAvatarGenerator` and `Program.cs` were added to the commit.

**CAUTION:**

It also said:

```
If running on Linux, you may need additional native dependencies (libgdiplus).
```

Signed-off-by: Tal Jacob <taljacob2@gmail.com>